go: split into multiple modules
