+++ .github/workflows/ci.yml
@@ -18,6 +18,7 @@ jobs:
           go-version: '>=1.19'
           check-latest: true
       - name: Run coverage
+        working-directory: ./pkg
         run: go test -coverprofile=coverage.out -covermode=count
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
