+++ .github/workflows/go.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       matrix:
         go-version: ['1.22']
-
+        dir: ['pkg', 'cmd/zstdseek']
     steps:
       - uses: actions/checkout@v4
       - name: Setup Go ${{ matrix.go-version }}
@@ -23,10 +23,13 @@ jobs:
           cache-dependency-path: go.sum
       - name: Display Go version
         run: go version
-      - name: Install dependencies
+      - name: Install dependencies (${{ matrix.dir }})
+        working-directory: ./${{ matrix.dir }}
         run: |
           go get .
-      - name: Build
+      - name: Build (${{ matrix.dir }})
+        working-directory: ./${{ matrix.dir }}
         run: go build -v ./...
-      - name: Test with the Go CLI
-        run: go test
+      - name: Test (${{ matrix.dir }})
+        working-directory: ./${{ matrix.dir }}
+        run: go test -v ./...
