+++ pkg/doc.go
