+++ pkg/encoder.go
