+++ pkg/encoder_test.go
