+++ pkg/reader_fuzz_test.go
