+++ pkg/seekable.go
