+++ pkg/seekable_fuzz_test.go
